+++ src/Castle.Core/DynamicProxy/IInvocation.cs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#nullable enable
+
 namespace Castle.DynamicProxy
 {
 	using System;
@@ -26,13 +28,13 @@ namespace Castle.DynamicProxy
 		///   Gets the arguments that the <see cref = "Method" /> has been invoked with.
 		/// </summary>
 		/// <value>The arguments the method was invoked with.</value>
-		object[] Arguments { get; }
+		object?[] Arguments { get; }
 
 		/// <summary>
 		///   Gets the generic arguments of the method.
 		/// </summary>
 		/// <value>The generic arguments, or null if not a generic method.</value>
-		Type[] GenericArguments { get; }
+		Type[]? GenericArguments { get; }
 
 		/// <summary>
 		///   Gets the object on which the invocation is performed. This is different from proxy object
@@ -40,7 +42,7 @@ namespace Castle.DynamicProxy
 		/// </summary>
 		/// <seealso cref = "IChangeProxyTarget" />
 		/// <value>The invocation target.</value>
-		object InvocationTarget { get; }
+		object? InvocationTarget { get; }
 
 		/// <summary>
 		///   Gets the <see cref = "MethodInfo" /> representing the method being invoked on the proxy.
@@ -52,7 +54,7 @@ namespace Castle.DynamicProxy
 		///   For interface proxies, this will point to the <see cref = "MethodInfo" /> on the target class.
 		/// </summary>
 		/// <value>The method invocation target.</value>
-		MethodInfo MethodInvocationTarget { get; }
+		MethodInfo? MethodInvocationTarget { get; }
 
 		/// <summary>
 		///   Gets the proxy object on which the intercepted method is invoked.
@@ -64,20 +66,20 @@ namespace Castle.DynamicProxy
 		///   Gets or sets the return value of the method.
 		/// </summary>
 		/// <value>The return value of the method.</value>
-		object ReturnValue { get; set; }
+		object? ReturnValue { get; set; }
 
 		/// <summary>
 		///   Gets the type of the target object for the intercepted method.
 		/// </summary>
 		/// <value>The type of the target object.</value>
-		Type TargetType { get; }
+		Type? TargetType { get; }
 
 		/// <summary>
 		///   Gets the value of the argument at the specified <paramref name = "index" />.
 		/// </summary>
 		/// <param name = "index">The index.</param>
 		/// <returns>The value of the argument at the specified <paramref name = "index" />.</returns>
-		object GetArgumentValue(int index);
+		object? GetArgumentValue(int index);
 
 		/// <summary>
 		///   Returns the concrete instantiation of the <see cref = "Method" /> on the proxy, with any generic
@@ -102,7 +104,7 @@ namespace Castle.DynamicProxy
 		/// <remarks>
 		///   In debug builds this can be slower than calling <see cref = "MethodInvocationTarget" />.
 		/// </remarks>
-		MethodInfo GetConcreteMethodInvocationTarget();
+		MethodInfo? GetConcreteMethodInvocationTarget();
 
 		/// <summary>
 		///   Proceeds the call to the next interceptor in line, and ultimately to the target method.
@@ -130,6 +132,6 @@ namespace Castle.DynamicProxy
 		/// </remarks>
 		/// <param name = "index">The index of the argument to override.</param>
 		/// <param name = "value">The new value for the argument.</param>
-		void SetArgumentValue(int index, object value);
+		void SetArgumentValue(int index, object? value);
 	}
 }